feat: refatorar envio de e-mails e adicionar suporte para templates de exército
