+++ src/main/config/usecases/execute-scheduled-task-factory.ts
@@ -1,9 +1,10 @@
 import { PrismaTaskLogRepository } from '@infra/repositories/task-log/task-log-repository';
 import { ExecuteScheduledTaskUseCase } from '@usecases/task-scheduler/execute-scheduled-task-use-case';
-import { myTaskRunner } from '@main/jobs/my-task';
+import { sendEmailJobUseCase } from '@usecases/email/send-email-job-use-case';
+
 
 const taskLogRepo = new PrismaTaskLogRepository();
-const executeTaskUseCase = new ExecuteScheduledTaskUseCase(taskLogRepo, myTaskRunner);
+const executeTaskUseCase = new ExecuteScheduledTaskUseCase(taskLogRepo, sendEmailJobUseCase);
 
 export function executeScheduledTask(taskName: string): Promise<void> {
   return executeTaskUseCase.execute(taskName);
