+++ src/utils/military/build-military-ott-email.ts
