+++ src/utils/military/build-military-stt-email.ts
